Changed distance function to return a floored int instead of a float.
Began Java ShellAI.

+++ clients/library/game.cpp
@@ -50,7 +50,7 @@ using std::string;
 using std::ofstream;
 
 //distance helper function
-int dist(float x1, float y1, float x2, float y2)
+int dist(int x1, int y1, int x2, int y2)
 {
   return static_cast<int>(floor(sqrt(pow(x1 - x2, 2) + pow(y1 - y2, 2))));
 }
